https://jira.jboss.org/jira/browse/JBMESSAGING-379 - Message Chunk and some other pieces

+++ src/main/org/jboss/messaging/core/journal/impl/JournalFile.java
@@ -51,11 +51,11 @@ public interface JournalFile
 
    void extendOffset(final int delta);
 
-   int getOffset();
+   long getOffset();
 
    int getOrderingID();
 
-   void setOffset(final int offset);
+   void setOffset(final long offset);
 
    SequentialFile getFile();
 }
